Remove duplicate logic
