+++ src/main/kotlin/com/ing/serialization/bfl/api/Api.kt
@@ -1,44 +1,27 @@
 package com.ing.serialization.bfl.api
 
 import com.ing.serialization.bfl.serde.BinaryFixedLengthInputDecoder
-import com.ing.serialization.bfl.serde.BinaryFixedLengthOutputEncoder
 import com.ing.serialization.bfl.serde.SerdeError
 import com.ing.serialization.bfl.serde.element.Layout
 import kotlinx.serialization.modules.EmptySerializersModule
 import kotlinx.serialization.modules.SerializersModule
-import kotlinx.serialization.serializer
 import kotlinx.serialization.serializerOrNull
 import java.io.ByteArrayInputStream
-import java.io.ByteArrayOutputStream
 import java.io.DataInputStream
-import java.io.DataOutputStream
 import kotlin.reflect.KClass
 
 fun <T : Any> serialize(data: T, serializersModule: SerializersModule = EmptySerializersModule): ByteArray {
     val serializer = serializersModule.serializerOrNull(data::class.java)
         ?: throw SerdeError.NoTopLevelSerializer(data::class)
 
-    return ByteArrayOutputStream().use { output ->
-        DataOutputStream(output).use { stream ->
-            BinaryFixedLengthOutputEncoder(stream, serializersModule).encodeSerializableValue(serializer, data)
-        }
-
-        output.toByteArray()
-    }
+    return genericSerialize(data, serializersModule, serializer)
 }
 
 fun <T : Any> debugSerialize(data: T, serializersModule: SerializersModule = EmptySerializersModule): Pair<ByteArray, Layout> {
     val serializer = serializersModule.serializerOrNull(data::class.java)
         ?: throw SerdeError.NoTopLevelSerializer(data::class)
 
-    return ByteArrayOutputStream().use { output ->
-        val layout = DataOutputStream(output).use { stream ->
-            val bfl = BinaryFixedLengthOutputEncoder(stream, serializersModule)
-            bfl.encodeSerializableValue(serializer, data)
-            bfl.layout
-        }
-        Pair(output.toByteArray(), layout)
-    }
+    return genericDebugSerialize(data, serializersModule, serializer)
 }
 
 fun <T : Any> deserialize(data: ByteArray, klass: KClass<T>, serializersModule: SerializersModule = EmptySerializersModule): T {
