+++ src/main/kotlin/com/ing/serialization/bfl/api/reified/Api.kt
@@ -1,37 +1,24 @@
 package com.ing.serialization.bfl.api.reified
 
+import com.ing.serialization.bfl.api.genericDebugSerialize
+import com.ing.serialization.bfl.api.genericSerialize
 import com.ing.serialization.bfl.serde.BinaryFixedLengthInputDecoder
-import com.ing.serialization.bfl.serde.BinaryFixedLengthOutputEncoder
 import com.ing.serialization.bfl.serde.element.Layout
 import kotlinx.serialization.modules.EmptySerializersModule
 import kotlinx.serialization.modules.SerializersModule
 import kotlinx.serialization.serializer
 import java.io.ByteArrayInputStream
-import java.io.ByteArrayOutputStream
 import java.io.DataInputStream
-import java.io.DataOutputStream
 
 inline fun <reified T : Any> serialize(data: T, serializersModule: SerializersModule = EmptySerializersModule): ByteArray {
     val serializer = serializersModule.serializer<T>()
-    return ByteArrayOutputStream().use { output ->
-        DataOutputStream(output).use { stream ->
-            BinaryFixedLengthOutputEncoder(stream, serializersModule).encodeSerializableValue(serializer, data)
-        }
-        output.toByteArray()
-    }
+    return genericSerialize(data, serializersModule, serializer)
 }
 
 inline fun <reified T : Any> debugSerialize(data: T, serializersModule: SerializersModule = EmptySerializersModule): Pair<ByteArray, Layout> {
     val serializer = serializersModule.serializer<T>()
 
-    return ByteArrayOutputStream().use { output ->
-        val layout = DataOutputStream(output).use { stream ->
-            val bfl = BinaryFixedLengthOutputEncoder(stream, serializersModule)
-            bfl.encodeSerializableValue(serializer, data)
-            bfl.layout
-        }
-        Pair(output.toByteArray(), layout)
-    }
+    return genericDebugSerialize(data, serializersModule, serializer)
 }
 
 inline fun <reified T : Any> deserialize(data: ByteArray, serializersModule: SerializersModule = EmptySerializersModule): T =
